Simplify style definitions of selections when using multiple selectors
Usually all selectable elements are children of a selection root with a
specific ID, so style definitions can use the ID as a distinction criteria.
The only exception is the selection visualisation as it is a direct child of
<body>. The selection class now explicitly allows multiple classes so that one
can e.g. have a common class for all selection elements and specific ones for
each selector instance.